// src/enums/postTag.ts
export enum PostTag {
  Android = 'android',
  Angular = 'angular',
  Bootstrap = 'bootstrap',
  CSharp = 'c-sharp',
  Jest = 'jest',
  MySQL = 'mysql',
  NetCore = 'net-core',
  NestJS = 'nest-js',
  NextJS = 'next-js',
  ReactJS = 'react-js',
  Redux = 'redux',
  SqlServer = 'sql-server',
  TypeScript = 'typescript',
  Windows10 = 'windows-10',
  WSL = 'wsl',
  Xamarin = 'xamarin',
  Yarn = 'yarn',
}